Fewer crashes
Was able to get rid of seemingly random crashes, game still crashes at the end if player does a "bad" move when there is a "good" move available. Example, if you hit the up arrow and it would combine two tiles but you hit left instead, instead of ending the game it crashes. Also, when you lose a game and hit reset button, game resets but you can also see "Game over" in the background.